Handle nulls in grouping aggs (ESQL-626)
This PR handles nulls in grouping keys and aggregating values.

Closes ESQL-525
Closes ESQL-634

+++ benchmarks/src/main/java/org/elasticsearch/benchmark/compute/operation/AggregatorBenchmark.java
@@ -126,7 +126,7 @@ public class AggregatorBenchmark {
                     long group = g;
                     long sum = LongStream.range(0, BLOCK_LENGTH).filter(l -> l % GROUPS == group).sum();
                     long count = LongStream.range(0, BLOCK_LENGTH).filter(l -> l % GROUPS == group).count();
-                    double expected = sum / count;
+                    double expected = (double) sum / count;
                     if (dValues.getDouble(g) != expected) {
                         throw new AssertionError(prefix + "expected [" + expected + "] but was [" + dValues.getDouble(g) + "]");
                     }
